// Frontend/src/phaser/ShooterPlatformer/interfaces/enums/EnemyEnums.ts
export enum enemyStates {
  enemy_idle = 'idle',
  enemy_walk = 'walk',
  enemy_dead = 'dead',
  enemy_alert = 'alert',
  enemy_hurt = 'hurt',
  enemy_aggresive = 'aggresive',
  enemy_attack = 'attack',
}

export enum enemyAnims {
  enemy_idle = 'idle',
  enemy_walk = 'walk',
  enemy_run = 'run',
  enemy_dead = 'dead',
  enemy_hurt = 'hurt',
  enemy_attack = 'attack',
}